+++ include/cutlass/gemm/device/gemv.h
@@ -75,6 +75,8 @@ public:
   static ComplexTransform const kTransformB = GemvKernel::kTransformB;
 
   static int const kThreadCount = GemvKernel::kThreadCount;
+  static int const kThreadsPerRow = GemvKernel::kThreadsPerRow;
+
   static int const kStages = GemvKernel::kStages;
 
   static int const kAlignmentA = GemvKernel::kAlignmentA;
@@ -106,8 +108,23 @@ public:
   }
 
   /// Computes the grid shape
-  static dim3 get_grid_shape(Arguments const &args) { 
-    return dim3((args.problem_size.row() + (kThreadCount - 1)) / kThreadCount, 1, args.batch_count % 65565);
+  static dim3 get_grid_shape(Arguments const &args, dim3 const &block) { 
+    if(platform::is_same<LayoutA, layout::ColumnMajor>::value) {
+      return dim3((args.problem_size.row() + (block.x - 1)) / block.x, 1, args.batch_count % 65536);
+    }
+    else {
+      return dim3((args.problem_size.row() + (block.y - 1)) / block.y, 1, args.batch_count % 65536);
+    }
+  }
+
+  /// Computes the block shape
+  static dim3 get_block_shape() { 
+    if(platform::is_same<LayoutA, layout::ColumnMajor>::value) {
+      return dim3(kThreadCount, 1, 1);
+    }
+    else {
+      return dim3(kThreadsPerRow, kThreadCount / kThreadsPerRow, 1);
+    }
   }
 
   /// Initializes Gemv state from arguments.
@@ -124,8 +141,8 @@ public:
   /// Runs the kernel using initialized state.
   Status run(cudaStream_t stream = nullptr) {
 
-    dim3 grid = get_grid_shape(params_);
-    dim3 block(GemvKernel::kThreadCount, 1, 1);
+    dim3 block = get_block_shape();
+    dim3 grid = get_grid_shape(params_, block);
 
     int smem_size = int(sizeof(typename GemvKernel::SharedStorage));
     
@@ -137,11 +154,7 @@ public:
     //
     cudaError_t result = cudaGetLastError();
 
-    if (result != cudaSuccess) {
-      return Status::kErrorInternal;
-    }
-  
-    return Status::kSuccess;
+    return result == cudaSuccess ? Status::kSuccess : Status::kErrorInternal;
   }
 
   /// Runs the kernel using initialized state.