+++ test/unit/gemm/device/gemv.cu
@@ -98,7 +98,7 @@ public:
     cutlass::Distribution::Kind init_A_ = cutlass::Distribution::Uniform,
     cutlass::Distribution::Kind init_B_ = cutlass::Distribution::Uniform,
     cutlass::Distribution::Kind init_C_ = cutlass::Distribution::Uniform,
-    uint64_t seed_ = 2080
+    uint64_t seed_ = 2023
   ):
     init_A(init_A_), init_B(init_B_), init_C(init_C_), seed(seed_) { }
 
@@ -156,22 +156,29 @@ public:
 
   /// Initializes data structures
   void initialize(
-    cutlass::MatrixCoord problem_size
+    cutlass::MatrixCoord problem_size,
+    int32_t batch_count
   ) {
 
     //
-    // Allocate the GEMM workspace
+    // Allocate the GEMV workspace
     //
 
-    tensor_A.resize(problem_size);
-    tensor_B.resize({problem_size.column(), 1});
-    tensor_C.resize({problem_size.row(), 1});
-    tensor_D.resize({problem_size.row(), 1});
-    reference_D.resize({problem_size.row(), 1}, false);
+    if(std::is_same<LayoutA, cutlass::layout::ColumnMajor>::value) {
+      tensor_A.resize({problem_size.row(), batch_count * problem_size.column()});
+    }
+    else {
+      tensor_A.resize({batch_count * problem_size.row(), problem_size.column()});
+    }
+    
+    tensor_B.resize({batch_count * problem_size.column(), 1});
+    tensor_C.resize({batch_count * problem_size.row(), 1});
+    tensor_D.resize({batch_count * problem_size.row(), 1});
+    reference_D.resize({batch_count * problem_size.row(), 1}, false);
 
-    EXPECT_TRUE(initialize_tensor(tensor_A.host_view(), init_A, seed + 2019));
-    EXPECT_TRUE(initialize_tensor(tensor_B.host_view(), init_B, seed + 2018));
-    EXPECT_TRUE(initialize_tensor(tensor_C.host_view(), init_C, seed + 2017));
+    EXPECT_TRUE(initialize_tensor(tensor_A.host_view(), init_A, seed + 1));
+    EXPECT_TRUE(initialize_tensor(tensor_B.host_view(), init_B, seed + 2));
+    EXPECT_TRUE(initialize_tensor(tensor_C.host_view(), init_C, seed + 3));
 
     // It is possible to randomly initialize to all zeros, so override this with non-zeros
     // in the upper left corner of each operand.
@@ -225,9 +232,14 @@ public:
     return passed;
   }
 
-  /// Verifies the result is a GEMM
+  /// Verifies the result
   bool verify(
-    cutlass::MatrixCoord problem_size, 
+    cutlass::MatrixCoord problem_size,
+    int32_t batch_count,
+    int64_t batch_stride_A,
+    int64_t batch_stride_B,
+    int64_t batch_stride_C,
+    int64_t batch_stride_D,
     ElementCompute alpha, 
     ElementCompute beta) {
 
@@ -242,7 +254,7 @@ public:
         ElementCompute, ElementAccumulator
     >(
       {problem_size.row(), 1, problem_size.column()},
-      alpha, 
+      alpha,
       tensor_A.host_ref(),
       Gemv::kTransformA,
       tensor_B.host_ref(),
@@ -250,7 +262,12 @@ public:
       beta, 
       tensor_C.host_ref(), 
       reference_D.host_ref(), 
-      ElementAccumulator(0)
+      ElementAccumulator(0),
+      batch_count,
+      batch_stride_A,
+      batch_stride_B,
+      batch_stride_C,
+      batch_stride_D
     );
 
     return compare_reference(problem_size, alpha, beta);
@@ -259,39 +276,50 @@ public:
   /// Runs one problem size
   bool run(
     cutlass::MatrixCoord problem_size, 
+    int32_t batch_count,
+    int64_t batch_stride_A,
+    int64_t batch_stride_B,
+    int64_t batch_stride_C,
+    int64_t batch_stride_D,
     ElementCompute alpha,
     ElementCompute beta) {
 
-    this->initialize(problem_size);
+    this->initialize(problem_size, batch_count);
 
     //
-    // Initialize the GEMM operator
+    // Initialize the GEMV operator
     //
 
     typename Gemv::Arguments arguments{
       problem_size,
+      batch_count,
       {alpha, beta},
       tensor_A.device_ref(),
       tensor_B.device_data(),
       tensor_C.device_data(),
       tensor_D.device_data(),
-      tensor_B.layout().stride(0),
-      tensor_C.layout().stride(0),
-      tensor_D.layout().stride(0)
+      batch_stride_A,
+      batch_stride_B,
+      batch_stride_C,
+      batch_stride_D
     };
 
     Gemv gemm_op;
+
+    cutlass::Status status = gemm_op.can_implement(arguments);
+
+    EXPECT_TRUE(status == cutlass::Status::kSuccess) << to_string(status);
 
     size_t workspace_size = Gemv::get_workspace_size(arguments);
 
     cutlass::device_memory::allocation<uint8_t> workspace(workspace_size);
 
-    cutlass::Status status = gemm_op.initialize(arguments, workspace.get());
+    status = gemm_op.initialize(arguments, workspace.get());
 
     EXPECT_TRUE(status == cutlass::Status::kSuccess) << to_string(status);
 
     //
-    // Run the GEMM
+    // Run the GEMV
     //
 
     status = gemm_op();
@@ -302,8 +330,15 @@ public:
     // Verify
     //
 
-    bool passed = this->verify(problem_size, alpha, beta);
-
+    bool passed = this->verify(
+        problem_size,
+        batch_count,
+        batch_stride_A,
+        batch_stride_B,
+        batch_stride_C,
+        batch_stride_D,
+        alpha,
+        beta);
     return passed;
   }
 };
@@ -315,12 +350,16 @@ bool TestAllGemv() {
 
   using ElementCompute = typename Gemv::EpilogueOutputOp::ElementCompute;
 
+  int Batch[] = {
+    1, 520, 1314
+  };
+
   int M[] = {
-    8, 48, 192, 520
+    1, 5, 16
   };
 
   int K[] = {
-    8, 192, 528
+    8, 128, 256
   };
 
   double Alpha[] = {
@@ -331,15 +370,25 @@ bool TestAllGemv() {
     0, 1, 1.25
   };
 
-  for (int m : M) {
-    for (int k : K) {
-      for (double alpha : Alpha) {
-        for (double beta : Beta) {
+  for (int b : Batch) {
+    for (int m : M) {
+      for (int k : K) {
+        for (double alpha : Alpha) {
+          for (double beta : Beta) {
 
-          TestbedGemv<Gemv> testbed;
+            TestbedGemv<Gemv> testbed;
 
-          if (!testbed.run({m, k}, ElementCompute(alpha), ElementCompute(beta))) {
-            return false;
+            if (!testbed.run(
+                    {m, k},
+                    b,
+                    m * k,
+                    k,
+                    m,
+                    m,
+                    ElementCompute(alpha),
+                    ElementCompute(beta))) {
+              return false;
+            }
           }
         }
       }
@@ -354,9 +403,103 @@ bool TestAllGemv() {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
-TEST(SM50_Device_Gemv_f32n_f32_f32_simt_f32, Simple) {
+TEST(SM50_Device_Gemv_f16n_f16_f16_simt_f32, RowMajorA) {
 
+  using ElementInput = cutlass::half_t;
+  using ElementOutput = cutlass::half_t;
+  using LayoutA = cutlass::layout::RowMajor;
+  using ElementAccumulator = float;
+  int const kElementsPerAccess = 8;
+  
+  using EpilogueOp = cutlass::epilogue::thread::LinearCombination<
+      ElementOutput,
+      1,
+      ElementAccumulator,
+      ElementAccumulator>;
+
+  using Gemv = cutlass::gemm::device::Gemv<
+      cutlass::gemm::kernel::Gemv<
+          ElementInput,           // Element A
+          LayoutA,                // Layout A
+          ElementInput,           // Element B
+          ElementOutput,          // Element C
+          ElementAccumulator,     // Element accumulator
+          EpilogueOp,             // Output operator
+          kElementsPerAccess      // Element access granularity
+          >
+      >;
+
+  EXPECT_TRUE(test::gemm::TestAllGemv<Gemv>());
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+TEST(SM50_Device_Gemv_f32n_f32_f32_simt_f32, RowMajorA) {
+
+  using ElementInput = float;
   using ElementOutput = float;
+  using LayoutA = cutlass::layout::RowMajor;
+  using ElementAccumulator = float;
+  int const kElementsPerAccess = 4;
+  
+  using EpilogueOp = cutlass::epilogue::thread::LinearCombination<
+      ElementOutput,
+      1,
+      ElementAccumulator,
+      ElementAccumulator>;
+
+  using Gemv = cutlass::gemm::device::Gemv<
+      cutlass::gemm::kernel::Gemv<
+          ElementInput,           // Element A
+          LayoutA,                // Layout A
+          ElementInput,           // Element B
+          ElementOutput,          // Element C
+          ElementAccumulator,     // Element accumulator
+          EpilogueOp,             // Output operator
+          kElementsPerAccess      // Element access granularity
+          >
+      >;
+
+  EXPECT_TRUE(test::gemm::TestAllGemv<Gemv>());
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+TEST(SM50_Device_Gemv_f64n_f64_f64_simt_f64, RowMajorA) {
+
+  using ElementInput = double;
+  using ElementOutput = double;
+  using LayoutA = cutlass::layout::RowMajor;
+  using ElementAccumulator = double;
+  int const kElementsPerAccess = 2;
+  
+  using EpilogueOp = cutlass::epilogue::thread::LinearCombination<
+      ElementOutput,
+      1,
+      ElementAccumulator,
+      ElementAccumulator>;
+
+  using Gemv = cutlass::gemm::device::Gemv<
+      cutlass::gemm::kernel::Gemv<
+          ElementInput,           // Element A
+          LayoutA,                // Layout A
+          ElementInput,           // Element B
+          ElementOutput,          // Element C
+          ElementAccumulator,     // Element accumulator
+          EpilogueOp,             // Output operator
+          kElementsPerAccess      // Element access granularity
+          >
+      >;
+
+  EXPECT_TRUE(test::gemm::TestAllGemv<Gemv>());
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+TEST(SM50_Device_Gemv_f16n_f16_f16_simt_f32, ColumnMajorA) {
+
+  using ElementInput = cutlass::half_t;
+  using ElementOutput = cutlass::half_t;
   using LayoutA = cutlass::layout::ColumnMajor;
   using ElementAccumulator = float;
 
@@ -368,9 +511,9 @@ TEST(SM50_Device_Gemv_f32n_f32_f32_simt_f32, Simple) {
 
   using Gemv = cutlass::gemm::device::Gemv<
     cutlass::gemm::kernel::Gemv<
-        ElementOutput,          // Element A
+        ElementInput,           // Element A
         LayoutA,                // Layout A
-        ElementOutput,          // Element B
+        ElementInput,           // Element B
         ElementOutput,          // Element C
         ElementAccumulator,     // Element Accumulator
         EpilogueOp              // Output operator
@@ -383,9 +526,9 @@ TEST(SM50_Device_Gemv_f32n_f32_f32_simt_f32, Simple) {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
-TEST(SM50_Device_Gemv_f16n_f16_f32_simt_f32, Simple) {
+TEST(SM50_Device_Gemv_f32n_f32_f32_simt_f32, ColumnMajorA) {
 
-  using ElementInput = cutlass::half_t;
+  using ElementInput = float;
   using ElementOutput = float;
   using LayoutA = cutlass::layout::ColumnMajor;
   using ElementAccumulator = float;
@@ -413,12 +556,12 @@ TEST(SM50_Device_Gemv_f16n_f16_f32_simt_f32, Simple) {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
-TEST(SM50_Device_Gemv_f16n_f16_f16_simt_f32, Simple) {
+TEST(SM50_Device_Gemv_f64n_f64_f64_simt_f64, ColumnMajorA) {
 
-  using ElementInput = cutlass::half_t;
-  using ElementOutput = cutlass::half_t;
+  using ElementInput = double;
+  using ElementOutput = double;
   using LayoutA = cutlass::layout::ColumnMajor;
-  using ElementAccumulator = float;
+  using ElementAccumulator = double;
 
   using EpilogueOp = cutlass::epilogue::thread::LinearCombination<
       ElementOutput,